+++ src/Service/StoreServiceImplementation.java
@@ -27,7 +27,10 @@ public class StoreServiceImplementation implements StoreService {
 
 	@Override
 	public Integer delete(String userName, String adminPassword) throws ApplicationErrorException {
-		return storeDAO.delete(userName, adminPassword);
+		if(userName != null && adminPassword != null)
+			return storeDAO.delete(userName, adminPassword);
+		else
+			throw new NullPointerException(">> Username Or Password Cannot be Null!!");
 	}
 
 	/**
@@ -36,6 +39,8 @@ public class StoreServiceImplementation implements StoreService {
 	 * @param store Store to be validated
 	 */
 	private void validate(Store store) throws InvalidTemplateException {
+		if(store == null)
+			throw new NullPointerException(">> Store cannot be Null!!");
 		if(store.getName() != null && ! store.getName().matches(NAME_REGEX))
 			throw new InvalidTemplateException(">> Invalid Store Name!!");
 		if(store.getPhoneNumber() != 0 && ! String.valueOf(store.getPhoneNumber()).matches(PHONE_NUMBER_REGEX))
